Some comments in the code

+++ video_fft/src/export.cpp
@@ -1,4 +1,4 @@
-int do_fft_compress(void *blob, int len_t, int len_y, int len_x, int len_c,
-                    float fps, float freq_limit, void *dest);
+int do_fft_compress(void *blob, uint16_t len_t, uint16_t len_y, uint16_t len_x,
+                    uint8_t len_c, float fps, float freq_limit, void *dest) ;
 
 int do_debug();
